requirements.txt
beautifulsoup4==4.6.0
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
google-api-core==1.21.0
google-auth==1.17.2
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.25.0
google-cloud-core==1.3.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
idna==2.6
numpy==1.14.1
oauthlib==3.1.0
pandas==0.22.0
pandas-gbq==0.13.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.1.0
python-dateutil==2.8.1
python-dotenv==0.8.2
pytz==2020.1
requests==2.18.4
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
six==1.15.0
tqdm==4.47.0
urllib3==1.25.9
xlrd==1.2.0
